requirements.txt
openai
pdfplumber
tiktoken
python-dotenv
